+++ cartographer/cartographer.cpp
@@ -22,9 +22,9 @@ using namespace llvm;
 
 bool noProgressBar;
 bool blocksLabeled = false;
-map<int, set<string>> blockLabelMap;
-map<int, BasicBlock *> blockMap;
-set<int> ValidBlocks;
+map<int64_t, set<string>> blockLabelMap;
+map<int64_t, BasicBlock *> blockMap;
+set<int64_t> ValidBlocks;
 
 llvm::cl::opt<string> inputTrace("i", llvm::cl::desc("Specify the input trace filename"), llvm::cl::value_desc("trace filename"));
 llvm::cl::opt<float> threshold("t", cl::desc("The threshold of block grouping required to complete a kernel."), llvm::cl::value_desc("float"), llvm::cl::init(0.9));
@@ -150,7 +150,7 @@ int main(int argc, char **argv)
         auto type35Kernels = TypeThree::Process(type4Kernels);
         spdlog::info("Detected " + to_string(type35Kernels.size()) + " type 3.5 kernels");
 
-        map<int, set<int>> finalResult;
+        map<int, set<int64_t>> finalResult;
         int j = 0;
         for (auto set : type35Kernels)
         {
